+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Using ChatGPT to generate Release Notes (only en and ru locales). You should set global setting `open_ai_api_key` to use it.
 * Added global settings. You can set it using new `alex settings set` command.
 
 ## 1.5.4